refactor: Rename `lib/plugins/aws/lib/setBucketName.js`
