+++ lib/plugins/aws/deploy-list.js
@@ -3,7 +3,7 @@
 const { log, writeText } = require('@serverless/utils/log');
 const validate = require('./lib/validate');
 const findAndGroupDeployments = require('./utils/findAndGroupDeployments');
-const setBucketName = require('./lib/setBucketName');
+const setBucketName = require('./lib/set-bucket-name');
 
 class AwsDeployList {
   constructor(serverless, options) {
