+++ lib/plugins/aws/deploy/index.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const writeServiceOutputs = require('../../../cli/write-service-outputs');
 const extendedValidate = require('./lib/extendedValidate');
-const setBucketName = require('../lib/setBucketName');
+const setBucketName = require('../lib/set-bucket-name');
 const checkForChanges = require('./lib/checkForChanges');
 const monitorStack = require('../lib/monitor-stack');
 const checkIfBucketExists = require('../lib/check-if-bucket-exists');
