+++ lib/plugins/aws/lib/set-bucket-name.js
