+++ lib/plugins/aws/rollback.js
@@ -1,6 +1,6 @@
 'use strict';
 const validate = require('./lib/validate');
-const setBucketName = require('./lib/setBucketName');
+const setBucketName = require('./lib/set-bucket-name');
 const updateStack = require('./lib/updateStack');
 const monitorStack = require('./lib/monitor-stack');
 const waitForChangeSetCreation = require('./lib/wait-for-change-set-creation');
