+++ test/unit/lib/plugins/aws/lib/set-bucket-name.test.js
